add canvas component
